refactored the code and fixed FD-DC task list bug.
