+++ src/main/services/features/claimantResponse/claimantResponseTasklistService/claimantResponseTasklistService.ts
@@ -13,10 +13,11 @@ export function getClaimantResponseTaskLists (claim: Claim, claimId: string, lng
   const lang = getLng(lng);
   const taskGroups : TaskList[] = [];
   taskGroups.push(buildHowDefendantRespondSection(claim, claimId, lang));
-  if(claim.isPartialAdmissionNotPaid()) {
+  if(claim.isPartialAdmissionNotPaid() || (claim.isFullDefence() && claim.isRejectAllOfClaimDispute())) {
     taskGroups.push(buildWhatToDoNextSection(claim, claimId, lang));
   }
-  if(claim.isPartialAdmissionPaid() || claim.isFullDefence()){
+  if(claim.isPartialAdmissionPaid() || (claim.isFullDefence() && !claim.isRejectAllOfClaimDispute()))
+  {
     taskGroups.push(buildYourResponseSection(claim, claimId, lang));
   }
   taskGroups.push(buildClaimantHearingRequirementsSection(claim, claimId, lang));
