+++ src/main/services/features/claimantResponse/responseSection/buildYourResponseSection.ts
@@ -8,42 +8,36 @@ import {changeLabel} from 'common/utils/checkYourAnswer/changeButton';
 import {YesNo, YesNoUpperCase} from 'form/models/yesNo';
 import {ClaimResponseStatus} from 'models/claimResponseStatus';
 import {RESPONSEFORNOTPAIDPAYIMMEDIATELY} from 'models/claimantResponse/checkAnswers';
-import { buildSummaryForPayByInstallments, buildSummaryForPayBySetDate } from '../checkAnswers/checkAnswersPayBySetDateOrInstallments';
 
-function buildFDDisputeTheClaimSummaryRows(claim: Claim, claimId: string, lang : string | unknown) {
+function buildFDDisputeTheClaimSummaryRows(claim: Claim, claimId: string, lang : string | unknown)  {
   const intentionToProceedHref = constructResponseUrlWithIdParams(claimId, CLAIMANT_RESPONSE_INTENTION_TO_PROCEED_URL);
   const intentionToProceed = claim.claimantResponse?.intentionToProceed?.option === YesNo.YES ? YesNoUpperCase.YES : YesNoUpperCase.NO;
-  return summaryRow(t('PAGES.CHECK_YOUR_ANSWER.PROCEED_WITH_CLAIM', {lang}), t(`COMMON.${intentionToProceed}`, {lang}), intentionToProceedHref, changeLabel(lang));
+  return summarySection({
+    title: t('PAGES.CHECK_YOUR_ANSWER.YOUR_RESPONSE', {lang}),
+    summaryRows: [
+      summaryRow(t('PAGES.CHECK_YOUR_ANSWER.PROCEED_WITH_CLAIM', {lang}), t(`COMMON.${intentionToProceed}`, {lang}), intentionToProceedHref, changeLabel(lang)),
+    ],
+  });
 }
 
 function buildPartAdmitPayImmediatelySummaryRows(claim: Claim, claimId: string, lang : string | unknown) {
   const partAdmitAcceptedHref = constructResponseUrlWithIdParams(claimId, CLAIMANT_RESPONSE_SETTLE_ADMITTED_CLAIM_URL);
   const selectedOption = claim?.claimantResponse?.hasPartAdmittedBeenAccepted?.option;
-  return summaryRow(t('PAGES.CHECK_YOUR_ANSWER.DO_YOU_ACCEPT_OR_REJECT_THE_DEFENDANTS_ADMISSION', { lang }), t(RESPONSEFORNOTPAIDPAYIMMEDIATELY[selectedOption], { lang }), partAdmitAcceptedHref, changeLabel(lang));
+  return summarySection({
+    title: t('PAGES.CHECK_YOUR_ANSWER.YOUR_RESPONSE', {lang}),
+    summaryRows: [
+      summaryRow(t('PAGES.CHECK_YOUR_ANSWER.DO_YOU_ACCEPT_OR_REJECT_THE_DEFENDANTS_ADMISSION', {lang}), t(RESPONSEFORNOTPAIDPAYIMMEDIATELY[selectedOption], {lang}), partAdmitAcceptedHref, changeLabel(lang)),
+    ],
+  });
 }
 
 export const buildYourResponseSection = (claim: Claim, claimId: string, lang: string | unknown): SummarySection => {
-  const yourResponseToClaimSection = summarySection({
-    title: t('PAGES.CHECK_YOUR_ANSWER.YOUR_RESPONSE', {lang}),
-    summaryRows: [],
-  });
-  const isSignSettlement = claim.isSignASettlementAgreement();
-  const isSignSettlementForPayBySetDate = isSignSettlement && (claim.isPAPaymentOptionByDate() || claim.isFAPaymentOptionBySetDate());
-  const isSignSettlementForPayByInstallments = isSignSettlement && (claim.isPAPaymentOptionInstallments() || claim.isFAPaymentOptionInstallments());
 
   if (claim.isFullDefence() && claim.isRejectAllOfClaimDispute()) {
-    yourResponseToClaimSection.summaryList.rows.push(buildFDDisputeTheClaimSummaryRows(claim, claimId, lang));
+    return buildFDDisputeTheClaimSummaryRows(claim, claimId, lang);
   }
 
   if (claim.isPartialAdmission() && claim?.responseStatus === ClaimResponseStatus.PA_NOT_PAID_PAY_IMMEDIATELY) {
-    yourResponseToClaimSection.summaryList.rows.push(buildPartAdmitPayImmediatelySummaryRows(claim, claimId, lang));
+    return buildPartAdmitPayImmediatelySummaryRows(claim, claimId, lang);
   }
-
-  if (isSignSettlementForPayBySetDate)
-    yourResponseToClaimSection.summaryList.rows.push(...buildSummaryForPayBySetDate(claim, claimId, lang));
-
-  if (isSignSettlementForPayByInstallments)
-    yourResponseToClaimSection.summaryList.rows.push(...buildSummaryForPayByInstallments(claim, claimId, lang));
-
-  return yourResponseToClaimSection;
 };
